Added command to convert framework import syntax
The "Modernize module import" command replaces a framework import line
like `#import <Foundation/Foundation.h>` with the module syntax
`@import Foundation;`. It only acts on umbrella header imports where
the name of the header is the same as the name of the framework.

Added a snippet for Appledoc-style block comments starting with `/*!`.